+++ spinta/cli/push.py
@@ -12,7 +12,6 @@ from typer import echo
 
 from spinta import exceptions, commands
 from spinta.auth import get_client_id_from_name, get_clients_path
-from spinta.backends.helpers import validate_and_return_transaction, validate_and_return_begin
 from spinta.cli.helpers.auth import require_auth
 from spinta.cli.helpers.data import ensure_data_dir
 from spinta.cli.helpers.errors import ErrorCounter
@@ -23,15 +22,14 @@ from spinta.cli.helpers.push.utils import extract_dependant_nodes, load_initial_
 from spinta.cli.helpers.push.read import read_rows
 from spinta.cli.helpers.push.state import init_push_state
 from spinta.cli.helpers.push.sync import sync_push_state
-from spinta.cli.helpers.store import prepare_manifest
+from spinta.cli.helpers.store import prepare_manifest, attach_backends, attach_keymaps
 from spinta.client import get_access_token
 from spinta.client import get_client_credentials
-from spinta.components import Action, Context, Store
+from spinta.components import Action
 from spinta.components import Config
 from spinta.components import Mode
 from spinta.core.context import configure_context
 from spinta.datasets.keymaps.sync import sync_keymap
-from spinta.manifests.components import Manifest
 from spinta.types.namespace import sort_models_by_ref_and_base
 from spinta.utils.units import tobytes
 from spinta.utils.units import toseconds
@@ -160,8 +158,8 @@ def push(
         auth_client = get_client_id_from_name(get_clients_path(config), auth_client)
         require_auth(context, auth_client)
 
-        _attach_backends(context, store, manifest)
-        _attach_keymaps(context, store)
+        attach_backends(context, store, manifest)
+        attach_keymaps(context, store)
         error_counter = ErrorCounter(max_count=max_error_count)
 
         models = commands.traverse_ns_models(context, ns, manifest, Action.SEARCH, dataset_=dataset, source_check=True)
@@ -239,24 +237,3 @@ def push(
 
         if error_counter.has_errors():
             raise Exit(code=1)
-
-
-def _attach_backends(context: Context, store: Store, manifest: Manifest) -> None:
-    context.attach('transaction', validate_and_return_transaction, context, manifest.backend)
-    backends = set()
-    for backend in store.backends.values():
-        backends.add(backend.name)
-        context.attach(f'transaction.{backend.name}', validate_and_return_begin, context, backend)
-    for backend in manifest.backends.values():
-        backends.add(backend.name)
-        context.attach(f'transaction.{backend.name}', validate_and_return_begin, context, backend)
-    for dataset_ in commands.get_datasets(context, manifest).values():
-        for resource in dataset_.resources.values():
-            if resource.backend and resource.backend.name not in backends:
-                backends.add(resource.backend.name)
-                context.attach(f'transaction.{resource.backend.name}', validate_and_return_begin, context, resource.backend)
-
-
-def _attach_keymaps(context: Context, store: Store) -> None:
-    for keymap in store.keymaps.values():
-        context.attach(f'keymap.{keymap.name}', lambda: keymap)